Renamed Activities Folder
New Refactor added.
