+++ app/src/main/res/layout/activity_create_new_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activities.CreateNewProfile">
+    tools:context=".Controllers.CreateNewProfile">
 
     <TextView
         android:id="@+id/TextView_title"
